Restructure Rust layout to ./rust/{src,tests/{unit,integration,e2e}}
Mirror the JS layout so source and tests live under per-language roots,
and split the existing inline `mod tests` block into external test files
grouped by suite (unit / integration / e2e).

- crate moves: rust/core/Cargo.toml → rust/Cargo.toml,
  rust/core/src/lib.rs → rust/src/lib.rs
- workspace member updated from `rust/core` to `rust`
- Cargo manifest declares three [[test]] entries that fold the per-folder
  modules under tests/{unit,integration,e2e}/mod.rs

The unit suite covers pure helpers (token coverage, statement extraction,
classification, support-ratio bounds). Integration exercises the
issue-35 / issue-41 semantic translation pipeline and doublet relations.
E2E calls the extern "C" surface that the WebAssembly bridge will host.
