+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["rust/core"]
+members = ["rust"]
 resolver = "2"
