+++ rust/tests/integration/mod.rs
@@ -1,0 +1,9 @@
+//! Integration tests for `meta_expression_core`.
+//!
+//! These combine the higher-level pipeline pieces — semantic-translation
+//! lookups, issue-35 doublet relations, and the lexical glossary fallback —
+//! to verify they operate together as advertised.
+
+mod issue35_translation;
+mod issue41_glossary;
+mod semantic_relations;
